remove autoqasm module
