--- docs/api/qbraid_qir.autoqasm.rst
@@ -1,9 +1,0 @@
-:orphan:
-
-qbraid_qir.autoqasm
-=====================
-
-.. automodule:: qbraid_qir.autoqasm
-   :members:
-   :undoc-members:
-   :show-inheritance:
