+++ docs/conf.py
@@ -41,7 +41,7 @@ extensions = [
 # set_type_checking_flag = True
 autodoc_member_order = "bysource"
 autoclass_content = "both"
-autodoc_mock_imports = ["cirq", "openqasm3", "autoqasm"]
+autodoc_mock_imports = ["cirq", "openqasm3"]
 napoleon_numpy_docstring = False
 todo_include_todos = True
 mathjax_path = "https://cdn.jsdelivr.net/npm/mathjax@2/MathJax.js?config=TeX-AMS-MML_HTMLorMML"
