--- qbraid_qir/autoqasm/__init__.py
@@ -1,34 +1,0 @@
-# Copyright (C) 2024 qBraid
-#
-# This file is part of the qBraid-SDK
-#
-# The qBraid-SDK is free software released under the GNU General Public License v3
-# or later. You can redistribute and/or modify it under the terms of the GPL v3.
-# See the LICENSE file in the project root or <https://www.gnu.org/licenses/gpl-3.0.html>.
-#
-# THERE IS NO WARRANTY for the qBraid-SDK, as per Section 15 of the GPL v3.
-
-"""
-Module containing AutoQASM QIR functionality.
-
-.. currentmodule:: qbraid_qir.autoqasm
-
-Functions
------------
-
-.. autosummary::
-   :toctree: ../stubs/
-
-   autoqasm_to_qir
-
-Exceptions
------------
-
-.. autosummary::
-   :toctree: ../stubs/
-
-   AutoQasmConversionError
-
-"""
-from .convert import autoqasm_to_qir
-from .exceptions import AutoQasmConversionError
